(PC-9515) move enum out of libs/algolia
