+++ src/features/cheatcodes/pages/AppComponents.tsx
@@ -16,8 +16,8 @@ import { AccordionItem, CallToAction } from 'features/offer/components'
 import { BeneficiaryCeilings } from 'features/profile/components/BeneficiaryCeilings'
 import { NonBeneficiaryHeader } from 'features/profile/components/NonBeneficiaryHeader'
 import { SelectionLabel } from 'features/search/atoms/SelectionLabel'
+import { CATEGORY_CRITERIA } from 'features/search/enums'
 import { mapCategoryToIcon } from 'libs/parsers'
-import { CATEGORY_CRITERIA } from 'libs/search'
 import { Badge } from 'ui/components/Badge'
 import { Banner, BannerType } from 'ui/components/Banner'
 import { ProgressBar } from 'ui/components/bars/ProgressBar'
