+++ src/features/favorites/atoms/Favorite.tsx
@@ -13,8 +13,8 @@ import { Credit } from 'features/home/services/useAvailableCredit'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { useDistance } from 'features/offer/components/useDistance'
 import { OfferImage } from 'features/search/atoms/OfferImage'
+import { CATEGORY_CRITERIA } from 'features/search/enums'
 import { formatToFrenchDate, getFavoriteDisplayPrice, parseCategory } from 'libs/parsers'
-import { CATEGORY_CRITERIA } from 'libs/search'
 import { AppButton } from 'ui/components/buttons/AppButton'
 import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 import { ColorsEnum, getSpacing, Spacer, Typo } from 'ui/theme'
