+++ src/features/search/atoms/Hit.tsx
@@ -7,9 +7,10 @@ import styled from 'styled-components/native'
 import { mergeOfferData } from 'features/home/atoms/OfferTile'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { useDistance } from 'features/offer/components/useDistance'
+import { CATEGORY_CRITERIA } from 'features/search/enums'
 import { analytics } from 'libs/analytics'
 import { formatDates, getDisplayPrice, parseCategory } from 'libs/parsers'
-import { SearchHit, CATEGORY_CRITERIA } from 'libs/search'
+import { SearchHit } from 'libs/search'
 import { ColorsEnum, getSpacing, Spacer, Typo } from 'ui/theme'
 import { ACTIVE_OPACITY } from 'ui/theme/colors'
