+++ src/features/search/sections/Category.tsx
@@ -3,10 +3,10 @@ import styled from 'styled-components/native'
 
 import { AccordionItem } from 'features/offer/components'
 import { SelectionLabel, TitleWithCount } from 'features/search/atoms'
+import { CATEGORY_CRITERIA } from 'features/search/enums'
 import { useStagedSearch } from 'features/search/pages/SearchWrapper'
 import { SectionTitle } from 'features/search/sections/titles'
 import { useLogFilterOnce } from 'features/search/utils/useLogFilterOnce'
-import { CATEGORY_CRITERIA } from 'libs/search'
 import { getSpacing } from 'ui/theme'
 
 // First we filter out the 'All' category
