+++ src/libs/algolia/enums/index.ts
@@ -1,3 +1,2 @@
-export * from './criteriaEnums'
 export * from './facetsEnums'
 export * from './radiusFiltersEnums'
